Add stub CLI commands to replace mario app
Why these changes are being introduced:
This application is a replacement of an existing app written in Go. We
want to ensure that all functionality of the old app (mario) is
rewritten into this  app. Stubbing out all of the expected CLI commands
from mario will provide clear initial documentation in the codebase of
what the desired end functionality should be for this rewrite, and will
allow us to confirm that we've completed replacement of the full mario
app functionality.

How this addresses that need:
* Adds stub CLI commands representing all CLI commands from the old
  mario app, with documentation and expected parameters but no
  functionality.
* Adds initial tests confirming that each command can be executed
  successfully.
* Updates the CLI to group all commands under a single main command, and
  updates existing tests to match.
* Also updates the setup.cfg file to add pydocstyle configuration.

Side effects of this change:
None

Relevant ticket(s):
* https://mitlibraries.atlassian.net/browse/RDI-124