Update implemented commands
Why these changes are being introduced:

While working on the next tasks of promoting indexes and ingesting
documents, it became clear that some of the functionality supporting the
existing commands needed to be refactored. In particular, retrieving
information from the cluster needed to be separated from formatting the
information for display so that the retrieval functions could also be
used for other commands.

How this addresses that need:
* Splits list_aliases and list_indexes functions in the opensearch
  module into two functions each, one to get the data and one to format
  it for display.
* Adds a function to get all of the aliases for an index and includes
  that in the get_formatted_indexes display output.
* Updates existing tests and fixtures to match changes and adds tests
  for new functionality.
* Adds docstrings and informational comments where needed.