Implement promote command
Why these changes are being introduced:

Promoting indexes to various aliases is an important part of how we
differentiate which source indexes get searched by which APIs/UIs. The
promote command needs to be able to promote an index to one or more
aliases, including the "primary" alias that always includes the current
primary indexes for each source. Promotion also needs to demote any
existing indexes for the given source from any aliases that source has
indexes in so ensure aliased indexes are current and there is never more
than one index for a given source in a give alias.

How this addresses that need:
* Adds two new interrelated functions to the opensearch module: one to
  identify all existing aliased indexes for a provided source, and one
  to promote an index to one or more aliases (including any existing
  aliases for the index's source) and demote any existing aliased
  indexes for the same source.
* Updates the promote cli command to call the opensearch module
  promote_index function and then display updated information about the
  state of all aliased indexes in the cluster.
* Adds tests and fixtures for all new functionality.
* Sets an initial config value for the "primary alias" in the config
  module. How and where this value gets stored may change in the future.

Side effects of this change:
Once deployed with the appropriate indexes promoted to the appropriate
aliases, APIs will be able to point to relevant alias(es) instead of
using index prefixes to determine which indexes to search.

Relevant ticket(s):
* https://mitlibraries.atlassian.net/browse/TIMX-77
* https://mitlibraries.atlassian.net/browse/TIMX-8