Implement delete CLI command
Why these changes are being introduced:

We need to be able to delete indexes. Because this is currently a manual
process, this implementation adds a user confirmation prompt that wasn't
present in the mario delete command. It can be overridden with a flag
for local development or if we ever decide to automate index deletion.

How this addresses that need:
* Adds a delete_index function to the opensearch module.
* Adds a helper function to confirm an action via user input.
* Updates the delete command in the cli module to get user confirmation
  if needed, call the delete_index function if confirmed or force is
  passed, and display useful result output.
* Adds a custom error to provide more readable error message when an
  index is not found in the cluster.
* Adds tests and fixtures for new functionality.

Relevant ticket(s):
* https://mitlibraries.atlassian.net/browse/TIMX-79