Add demote command to CLI
Why these changes are being introduced:
Very occasionally we need to be able to demote an index from alias(es)
without promoting a different one for the source. This adds that
functionality.

How this addresses that need:
* Adds a remove_alias function to the opensearch module that removes a
  single alias from an index.
* Adds a demote command to the cli module that demotes an index from all
  its associated aliases, with a manual confirmation if the primary
  aliases is included (since we generally do not want to demote an index
  from the primary alias without promoting a different one for the
  source).
* Adds a custom AliasNotFoundError and updates error handling across
  several functions to distinguish between index vs. alias not found
  errors, which are handled identically by opensearchpy.
* Updates output formatting for the get_formatted_aliases function.
* Adds/updates tests and fixtures for all changes.

Side effects of this change:
It will now be possible to demote a source index from the primary alias
without promoting another index for the source, meaning the primary
alias could end up in a state where it doesn't include an index for each
source present in the cluster. The confirmation should help prevent this
being done accidentally, but we still need to be sure to proceed with
caution when manually demoting.

Relevant ticket(s):
* https://mitlibraries.atlassian.net/browse/TIMX-80