Only skip known indexing errors
Why these changes are being introduced:
Previously when bulk-indexing returned an error from Opensearch, we
logged the error, skipped the record, and continued indexing. We still
want to do that for the known error around mapper parsing (since that is
a data quality/transformation issue, not an Opensearch issue) but for
all other errors we want to stop the process with an exception for
further investigation.

How this addresses that need:
* Adds BulkIndexingError to errors module.
* During bulk indexing, checks for the known error we want to skip and
  otherwise raises the new exception.
* Updates tests and fixtures to reflect changes.
* Minor documentation update in cli module.

Side effects of this change:
If we hit any errors other than the occasional expected mapping error,
the Step Function will fail. This is good.