Support 'opensearch' as local host variation
Why these changes are being introduced:
Previously, only the string 'localhost' would trigger the code path to setup an Opensearch connection
to a locally running instance.  This proved problematic when attempting to have a docker container (e.g.
this TIM project) connect to another docker container (e.g. an Opensearch instance) because 'localhost'
was pointing at the calling TIM container and could not see the other.

By allowing the string 'opensearch' to also trigger the code path for a local connection, setting
the host to 'opensearch', this allows using docker networks and/or container names where one container
can make requests to opensearch on the host 'opensearch'.

How this addresses that need:
* both 'localhost' and 'opensearch' trigger a non SSL, non authenticated opensearch connection
* this TIM project running as a container can access an Opensearch client on the 'opensearch' host

Side effects of this change:
* None; the simple string 'opensearch' would not resolve to an actual, AWS deployed Opensearch instance

Relevant ticket(s):
* None