Remove parquet dataset feature flags
Why these changes are being introduced:

During development work of establishing the parquet dataset
architecture for TIMDEX ETL, feature flags were used in
various applications to remain backwards compatible.  Now
that the parquet dataset architecture is established, the
feature flags are no longer needed.

How this addresses that need:
* Removes 'bulk-index' and 'bulk-delete' CLI commands, preferring
only 'bulk-update' which expects a subset of records from the parquet
dataset to index and delete (per the 'action' column in the dataset).

Side effects of this change:
* Indexing from a JSON array of TIMDEX Record objects is no longer
supported, neither is a deleting from a JSON array of objects
with a single 'timdex_record_id' field.  All indexing and deleting
is expected to originate from parquet dataset records.

Relevant ticket(s):
* https://mitlibraries.atlassian.net/browse/TIMX-491