Add new reindex-source CLI command
Why these changes are being introduced:

Now that the TDA library can yield only the current records for a source from
the parquet dataset, this opens up the opporunity for a fairly simple CLI command
that performs a full refresh of a source in Opensearch using data already present
in the ETL dataset data, no extract or transform work is required.

This command could be helpful for development, where it would be handy to see
the current state of a source in a local Opensearch instance.  Or, it could be
helpful for disaster recovery where we want to refresh a source in Opensearch
without performing the extract + transform stages again.

How this addresses that need:
* Adds new reindex-source CLI command
  * create new index
  * promote and alias the new index
  * read current records for a source from dataset and bulk index them

Side effects of this change:
* None

Relevant ticket(s):
* https://mitlibraries.atlassian.net/browse/TIMX-495